Both project and package details are rendered in the same window
