+++ src/windows/package/package.html
@@ -1,38 +1,22 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<meta charset="UTF-8">
-	<title>Conductor - the Composer UI</title>
-	<link href="../../../node_modules/font-awesome/css/font-awesome.css" rel="stylesheet" type="text/css"/>
-	<link href="../../../public/css/conductor.css" rel="stylesheet" type="text/css"/>
-</head>
-<body>
-<div class="flex flex--column flex--grow" style="padding: 2rem">
-	<div class="flex flex--row package__title">
-		<h1 id="package-name"></h1>
-	</div>
-	<ul class="package__details flex flex--column">
-		<li class="package__details--item">
-			<span class="package__details--label">Version</span>
-			<span class="package__details--value" id="package-version"></span>
-		</li>
-		<li class="package__details--item">
-			<span class="package__details--label">Homepage</span>
-			<span class="package__details--value" id="package-homepage"></span>
-		</li>
-		<li class="package__details--item">
-			<span class="package__details--value" id="package-description"></span>
-		</li>
-	</ul>
-	<div class="flex flex--row button__actions">
-		<button class="flex" id="action-composer-update">Update <i class="fa fa-spin fa-circle-o-notch hidden"></i></button>
-		<button class="flex" id="action-composer-remove">Remove <i class="fa fa-spin fa-circle-o-notch hidden"></i></button>
-		<button class="flex" id="action-composer-show">Show <i class="fa fa-spin fa-circle-o-notch hidden"></i></button>
-	</div>
-	<div class="project__output" id="composer-output"></div>
+<div class="flex flex--row package__title">
+	<h1 id="package-name"></h1>
 </div>
-</body>
-<script>
-	require('./packageWindow.js')
-</script>
-</html>
+<ul class="package__details flex flex--column">
+	<li class="package__details--item">
+		<span class="package__details--label">Version</span>
+		<span class="package__details--value" id="package-version"></span>
+	</li>
+	<li class="package__details--item">
+		<span class="package__details--label">Homepage</span>
+		<span class="package__details--value" id="package-homepage"></span>
+	</li>
+	<li class="package__details--item">
+		<span class="package__details--value" id="package-description"></span>
+	</li>
+</ul>
+<div class="flex flex--row button__actions">
+	<button class="flex" id="action-composer-update">Update <i class="fa fa-spin fa-circle-o-notch hidden"></i></button>
+	<button class="flex" id="action-composer-remove">Remove <i class="fa fa-spin fa-circle-o-notch hidden"></i></button>
+	<button class="flex" id="action-composer-show">Show <i class="fa fa-spin fa-circle-o-notch hidden"></i></button>
+</div>
+<div class="project__output" id="composer-output"></div>
